+++ src/notelist.ts
@@ -9,6 +9,8 @@ import { ThumbnailCache, Thumbnail } from "./type";
 import * as naturalCompare from "string-natural-compare";
 import * as fs from "fs-extra";
 import notelistLogging from "electron-log/main";
+import { MenuItemLocation } from "api/types";
+import { ModelType } from "api/types";
 
 let i18n: any;
 
@@ -38,6 +40,7 @@ class Notelist {
     await this.genItemTemplate();
     await this.registerRendererPreview();
     await this.createMsgDialog();
+    await this.registerCommands();
   }
 
   private async setupLog() {
@@ -100,6 +103,14 @@ class Notelist {
 
     this.dataDir = await joplin.plugins.dataDir();
 
+    let confidentialTags = await joplin.settings.value("confidentialTags");
+    confidentialTags = confidentialTags.trim().toLowerCase();
+    if (confidentialTags.indexOf(",") != 0 && confidentialTags.length > 0) {
+      confidentialTags = confidentialTags.split(/\s*,\s*/);
+    } else {
+      confidentialTags = [];
+    }
+
     this.settings = {
       itemTemplate: "",
       layout: await joplin.settings.value("layout"),
@@ -120,7 +131,58 @@ class Notelist {
       todoDueColorOverdue: await joplin.settings.value("todoDueColorOverdue"),
       todoDueColorDone: await joplin.settings.value("todoDueColorDone"),
       joplinZoome: await joplin.settings.globalValue("windowContentZoomFactor"),
+      confidentialTags: confidentialTags,
     };
+  }
+
+  private async registerCommands() {
+    await joplin.commands.register({
+      name: "toggleConfidential",
+      label: i18n.__("cmd.toggleConfidential"),
+      execute: async () => {
+        await this.toggleConfidential();
+      },
+    });
+
+    await joplin.views.menuItems.create(
+      "ToolsToggleConfidential",
+      "toggleConfidential",
+      MenuItemLocation.Tools
+    );
+    await joplin.views.menuItems.create(
+      "NoteListContextMenuToggleConfidential",
+      "toggleConfidential",
+      MenuItemLocation.NoteListContextMenu
+    );
+  }
+
+  private async toggleConfidential() {
+    const ids = await joplin.workspace.selectedNoteIds();
+    if (ids.length > 0) {
+      for (const noteId of ids) {
+        const confidential = await joplin.data.userDataGet(
+          ModelType.Note,
+          noteId,
+          "confidential"
+        );
+
+        let confidentialNew = true;
+        if (confidential) {
+          confidentialNew = false;
+        }
+
+        await joplin.data.userDataSet(
+          ModelType.Note,
+          noteId,
+          "confidential",
+          confidentialNew
+        );
+
+        this.log.verbose(
+          "Toggle confidential of note " + noteId + " to " + confidential
+        );
+      }
+    }
   }
 
   private async genItemTemplate(): Promise<void> {
@@ -493,7 +555,11 @@ class Notelist {
     }
   }
 
-  private async getFieldValue(field: string, props: any): Promise<string> {
+  private async getFieldValue(
+    field: string,
+    props: any,
+    confidential: boolean
+  ): Promise<string> {
     let value = " ";
     switch (field.toLowerCase()) {
       case "createdtime":
@@ -548,14 +614,18 @@ class Notelist {
         }
         break;
       case "notetext":
-        value = "{{noteBody}}";
+        if (confidential) {
+          value = "{{noteBody}}";
+        } else {
+          value = "{{noteBody}}";
+        }
         break;
       case "url":
         value = '<span class="url">' + props.note.source_url + "</span>";
         break;
       case "tags":
         const tags = await this.getTags(props.note.tags);
-        if (tags.length > 0) {
+        if (!confidential && tags.length > 0) {
           value =
             '<span class="tags"><span class="tag">' +
             tags.join('</span> <span class="tag">') +
@@ -573,7 +643,8 @@ class Notelist {
 
   private async replaceFieldPlaceholder(
     text: string,
-    noteFields: string[]
+    noteFields: string[],
+    confidential: boolean
   ): Promise<string> {
     // asyncStringReplace copied from https://dev.to/ycmjason/stringprototypereplace-asynchronously-28k9
     const asyncStringReplace = async (
@@ -598,7 +669,7 @@ class Notelist {
         text,
         /{{([^}]+)}}/g,
         async (match, groups) => {
-          return await this.getFieldValue(groups, noteFields);
+          return await this.getFieldValue(groups, noteFields, confidential);
         }
       );
     } catch (error) {
@@ -626,6 +697,33 @@ class Notelist {
     return returnValue;
   }
 
+  private async isNoteConfidential(
+    noteId: string,
+    tags: any
+  ): Promise<boolean> {
+    const confidential = await joplin.data.userDataGet(
+      ModelType.Note,
+      noteId,
+      "confidential"
+    );
+
+    if (confidential) {
+      return true;
+    }
+
+    if (this.settings["confidentialTags"].length > 0) {
+      for (const tag of tags) {
+        if (
+          this.settings["confidentialTags"].indexOf(tag.title.toLowerCase()) !==
+          -1
+        ) {
+          return true;
+        }
+      }
+    }
+    return false;
+  }
+
   private async onRenderNoteCall(props: any): Promise<any> {
     this.log.verbose("Func: onRenderNoteCall " + props.note.id);
     const start = performance.now();
@@ -633,6 +731,10 @@ class Notelist {
     const completed =
       props.note.is_todo == 1 && props.note.todo_completed != 0 ? true : false;
 
+    const confidential = await this.isNoteConfidential(
+      props.note.id,
+      props.note.tags
+    );
     let noteLine = this.settings["noteLine"];
     let firstLine = this.settings["firstLine"];
     let lastLine = this.settings["lastLine"];
@@ -640,22 +742,34 @@ class Notelist {
 
     let noteTextLeft = await this.replaceFieldPlaceholder(
       noteTextParts["noteTextLeft"],
-      props
+      props,
+      confidential
     );
     let noteTextRight = await this.replaceFieldPlaceholder(
       noteTextParts["noteTextRight"],
-      props
+      props,
+      confidential
     );
     let noteBody = "";
-    if (noteTextParts["noteText"] == true) {
+    if (!confidential && noteTextParts["noteText"] == true) {
       noteBody = await this.getBody(props.note.body);
+    } else if (confidential) {
+      noteBody = i18n.__("confidentialNotetext");
     }
 
-    firstLine = await this.replaceFieldPlaceholder(firstLine, props);
-    lastLine = await this.replaceFieldPlaceholder(lastLine, props);
+    firstLine = await this.replaceFieldPlaceholder(
+      firstLine,
+      props,
+      confidential
+    );
+    lastLine = await this.replaceFieldPlaceholder(
+      lastLine,
+      props,
+      confidential
+    );
 
     let thumbnail = null;
-    if (this.settings["thumbnail"] != "no") {
+    if (!confidential && this.settings["thumbnail"] != "no") {
       thumbnail = await this.getResourcePreview(props.note.id, props.note.body);
     }