+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## not released
 
 - Fix: Increase thumbnail size for better previews, the Joplin zoom level is also taken into account #18
+- Add: Option to mark notes as confidential #17
 
 ## v0.4.0 (2024-01-25)